:recycle: Remove react-loadable, refactor routes
