+++ src/views/NotFound.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const NotFound = () => <div>404</div>;
+
+export default NotFound;
